+++ src/main/java/net/mohron/skyclaims/command/argument/IslandArgument.java
@@ -26,7 +26,6 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
-import net.mohron.skyclaims.SkyClaims;
 import net.mohron.skyclaims.permissions.Permissions;
 import net.mohron.skyclaims.team.PrivilegeType;
 import net.mohron.skyclaims.world.Island;
@@ -65,8 +64,7 @@ public class IslandArgument extends CommandElement {
       if (IslandManager.ISLANDS.containsKey(uuid)) {
         return Sets.newHashSet(uuid);
       }
-    } catch (IllegalArgumentException e) {
-      SkyClaims.getInstance().getLogger().error("Error parsing Island argument:" , e);
+    } catch (IllegalArgumentException ignored) {
     }
     Set<UUID> islands = IslandManager.ISLANDS.entrySet().stream()
         .filter(i -> i.getValue().getOwnerName().equalsIgnoreCase(arg))