+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Removed outdated schematics:
   - Garden of Glass
   - SkyExchange
+- Added `World > Regen-On-Create` config option
 - Fixed schematics sometimes not generating at the intended height. The height set will be the height the player is at when standing on the lowest block of a schematic.
 - Fixed Nucleus Integration commands not registering after a reload
 - Fixed admin island expansion (`/is info`) bug where old clickable text can be used to expand outside the region